Add files via upload
game simulation. need to have a starting amount and an end amount. the program will run until the start amount equals the end amount or reaches zero. once it reaches one of those two parameters it will add one to a variable  called wins. it will repeat this process until it has reached a predefined number of wins.